+++ aim/cli/convert/processors/tensorboard.py
@@ -4,7 +4,7 @@ import os
 import click
 from tqdm import tqdm
 
-from aim import Image, Run
+from aim import Audio, Image, Run
 
 
 def parse_tb_logs(tb_logs, repo_inst, flat=False, no_cache=False):
@@ -208,6 +208,7 @@ def parse_tb_logs(tb_logs, repo_inst, flat=False, no_cache=False):
                         track_val = None
                         try:
                             if value.HasField('tensor'):
+                                # TODO: [MV] check the case when audios are passed via tensor
                                 if plugin_name == 'images':
                                     tensor = value.tensor.string_val[2:]
                                     track_val = [
@@ -221,6 +222,9 @@ def parse_tb_logs(tb_logs, repo_inst, flat=False, no_cache=False):
                                 track_val = value.simple_value
                             elif value.HasField('image'):
                                 track_val = Image(tf.image.decode_image(value.image.encoded_image_string).numpy())
+                            elif value.HasField('audio'):
+                                tf_audio, sample_rate = tf.audio.decode_wav(value.audio.encoded_audio_string)
+                                track_val = Audio(tf_audio.numpy(), sample_rate=sample_rate)
 
                         except RuntimeError as exc:
                             # catch all the nasty failures